requirements.txt
Django==4.2
djangorestframework==3.14
psycopg2-binary==2.9.9
gunicorn==21.2.0
requests==2.31.0
django-filter==23.5
djangorestframework-simplejwt==5.2.2
drf_spectacular
whitenoise